Create unmoved, which works like a jQuery plugin.
This plugin requires both jQuery and jQuery UI to be included. It loads
data from the Node.js server, moving objects into position. It also
attaches events to said objects, recording movements and sending them to
the server.

To use, include the script, and call the unmoved() method on a jQuery
object, setting the server location. e.g.

    $('.unmoved-moveable').unmoved({
      'objectsResource' : 'http://localhost:1137/pages/' + encodeURIComponent(document.URL) + '/objects'
    });