[#4] Only restrict access if REMOTE_HOST set.
Previously, access was restricted to REMOTE_HOST, regardless of the
environment, and REMOTE_HOST had to be set. This approach complicates
development, however. Also, although it is by no means recommended
that the server be run in production without access being restricted,
this should be an environment decision.

Change the server to only restrict access if REMOTE_HOST is set, and
change `.env.example` accordingly. Note that if REMOTE_HOST is set to
an empty string, then it will be treated as being set.