[#9] Simplify unmoved.js options, implementing host.
Previously, it was necessary to set `objectsResource` to the full path
of the resource, also including the encoded URL for the remote page.

Replace `objectsResource` option with `host`, moving the building of the
full URL into `src/unmoved.js` itself. Also set the `unmoved.js` default
host to be `''`, meaning that no host needs to be specified, for
development.

Update the demos to work with the new parameters, explicitly setting the
host, by way of example.

Update the documentation, to reflect the new usage.