[#18] Create chess demo.
The chess demo demonstrates the simplicity with which more complex
things can be built out of unmoved. There is no server operating for the
game, besides the normal unmoved pages_objects server. A new game is
created by visiting a random URL. Because unmoved stores the query-string
as part of the key, this is enough to reinitialise pieces. Anyone
visiting the same URL will be able to move the pieces, although they
will have to refresh.

It is, of course, possible to play chess using this demo. But that is
not its purpose; there are no rules, no move-tracking, no notifications,
etc. If people want to cheat, or knock the pieces over, they can. ;)